updates to vpnd & xuap to check for breaking pkgs

+++ xbps-update-all-possible.sh
@@ -20,22 +20,35 @@
 
 
         # Check if error (2>&1 redirects stderr>stdout)
-        /usr/bin/xbps-install -Munv 2>&1 | grep MISSING > /tmp/void-errs
-        # if non-null then errors
+	/usr/bin/xbps-install -Munv 2>&1 | grep MISSING > /tmp/void-errs   # check for missing packages
+	/usr/bin/xbps-install -Munv 1>/dev/null 2>/tmp/stderr   # need to do in 2 steps for some reason
+	grep broken /tmp/stderr > /tmp/void-breaks   # check for broken packages and breaking packages
+	# if non-null then errors
 	en=`cat /tmp/void-errs | /usr/bin/wc -l`
-	if [ $en -eq 0 ] # no error
+	eb=`cat /tmp/void-breaks | /usr/bin/wc -l`
+	if [[ $en -eq 0 && $eb -eq 0 ]]  # no error
 	then 
 	   sudo xbps-install -Su # just run regular update
         else # error
-	    rm /tmp/dontinstall
-	    while read line; do  # check each problematic pkg
+	    if [ -f /tmp/dontinstall ]
+	    then
+		rm /tmp/dontinstall
+	    fi
+	    while read line; do  # check each problematic pkg in void-errs
 		echo $line > /tmp/tmpline  # strip out each line about problem pkgs to tmp file /tmp/tmpline
 		problem=$(cut -d' ' -f2- /tmp/tmpline | sed 's/[<>=].*//') # narrow to pkg-name in /tmp/tmpline, store in $problem
 		echo "$(/usr/bin/xbps-query -X $problem | sed 's/-[0-9].*$//g')" >> /tmp/dontinstall # store list of all unupgradeable pkgs
 	    done < /tmp/void-errs
+
+    	    while read line; do  # check each problematic pkg in void-breaks
+		echo $line > /tmp/tmpline  # strip out each line about problem pkgs to tmp file /tmp/tmpline
+		problem=$(cut -d " " -f 1 /tmp/tmpline | sed 's/-[0-9].*:$//g') # narrow to pkg-name by throwing away final -DDD:
+#		echo $problem # DELETEME
+		echo $problem >> /tmp/dontinstall
+	    done < /tmp/void-breaks
    
 	    # Cut out relevant info from all lines with "Found"
-            /usr/bin/xbps-install -Munv 2>&1 |grep Found | cut -d' ' -f2 | sed 's/-[0-9].*$//g' > /tmp/void-pkgs 
+            /usr/bin/xbps-install -Munv 2>&1 |grep Found | cut -d' ' -f2 | cut -d' ' -f1 | sed 's/-[0-9].*$//g' > /tmp/void-pkgs 
 	    
 	    # Create list of all installable pkgs
 	    while read targetpkg; do
@@ -48,12 +61,22 @@
 			bad=$((bad+1))
 		    fi
 		done < /tmp/dontinstall
-		if [ $bad -eq 0 ]
+#		echo $bad
+#		echo $targetpkg
+		if [ $bad -eq 0 ]   # if targetpkg doesn't match a pkg already in our don't install list then check
 		then
-		    installme="$installme $targetpkg"
-#		    echo $installme
+		    sudo xbps-install -Sun $targetpkg 2>&1 | grep breaks > /tmp/morebreaks # check if pkg breaks anything
+    		    sudo xbps-install -Sun $targetpkg 2>&1 | grep broken >> /tmp/morebreaks # check if pkg breaks anything
+	    	    emb=`cat /tmp/morebreaks | /usr/bin/wc -l`
+	    	    if [ $emb -eq 0 ]   # if targetpkg doesn't break anything, then add to our install string
+		    then
+			
+			installme="$installme $targetpkg"
+		    fi
+#		    echo "installme="$installme
 		fi
 	    done < /tmp/void-pkgs
-	    sudo xbps-install -Su $installme
+	    sudo xbps-install -S $installme     # finally install everything that can be installed
+		
 	fi
